chore: refactored store and services for views
