+++ apps/web/ce/services/project/index.ts
@@ -1,2 +1,1 @@
 export * from "./estimate.service";
-export * from "./view.service";
