+++ apps/web/core/store/project-view.store.ts
@@ -2,12 +2,12 @@ import { set } from "lodash";
 import { observable, action, makeObservable, runInAction, computed } from "mobx";
 import { computedFn } from "mobx-utils";
 // types
-import { EViewAccess, IProjectView, TPublishViewDetails, TPublishViewSettings, TViewFilters } from "@plane/types";
+import { IProjectView, TViewFilters } from "@plane/types";
 // constants
 // helpers
 import { getValidatedViewFilters, getViewName, orderViews, shouldFilterView } from "@plane/utils";
 // services
-import { ViewService } from "@/plane-web/services";
+import { ViewService } from "@/services/view.service";
 // store
 import type { CoreRootStore } from "./root.store";
 
@@ -41,25 +41,6 @@ export interface IProjectViewStore {
   // favorites actions
   addViewToFavorites: (workspaceSlug: string, projectId: string, viewId: string) => Promise<any>;
   removeViewFromFavorites: (workspaceSlug: string, projectId: string, viewId: string) => Promise<any>;
-  // publish
-  publishView: (
-    workspaceSlug: string,
-    projectId: string,
-    viewId: string,
-    data: TPublishViewSettings
-  ) => Promise<TPublishViewDetails | undefined>;
-  fetchPublishDetails: (
-    workspaceSlug: string,
-    projectId: string,
-    viewId: string
-  ) => Promise<TPublishViewDetails | undefined>;
-  updatePublishedView: (
-    workspaceSlug: string,
-    projectId: string,
-    viewId: string,
-    data: Partial<TPublishViewSettings>
-  ) => Promise<void>;
-  unPublishView: (workspaceSlug: string, projectId: string, viewId: string) => Promise<void>;
 }
 
 export class ProjectViewStore implements IProjectViewStore {
@@ -87,8 +68,6 @@ export class ProjectViewStore implements IProjectViewStore {
       fetchViews: action,
       fetchViewDetails: action,
       // CRUD actions
-      createView: action,
-      updateView: action,
       deleteView: action,
       // actions
       updateFilters: action,
@@ -213,19 +192,17 @@ export class ProjectViewStore implements IProjectViewStore {
    * @param data
    * @returns Promise<IProjectView>
    */
-  createView = async (workspaceSlug: string, projectId: string, data: Partial<IProjectView>): Promise<IProjectView> => {
-    const response = await this.viewService.createView(workspaceSlug, projectId, getValidatedViewFilters(data));
+  createView = action(
+    async (workspaceSlug: string, projectId: string, data: Partial<IProjectView>): Promise<IProjectView> => {
+      const response = await this.viewService.createView(workspaceSlug, projectId, getValidatedViewFilters(data));
 
-    runInAction(() => {
-      set(this.viewMap, [response.id], response);
-    });
+      runInAction(() => {
+        set(this.viewMap, [response.id], response);
+      });
 
-    if (data.access === EViewAccess.PRIVATE) {
-      await this.updateViewAccess(workspaceSlug, projectId, response.id, EViewAccess.PRIVATE);
+      return response;
     }
-
-    return response;
-  };
+  );
 
   /**
    * Updates a view details of specific view and updates it in the store
@@ -235,29 +212,27 @@ export class ProjectViewStore implements IProjectViewStore {
    * @param data
    * @returns Promise<IProjectView>
    */
-  updateView = async (
-    workspaceSlug: string,
-    projectId: string,
-    viewId: string,
-    data: Partial<IProjectView>
-  ): Promise<IProjectView> => {
-    const currentView = this.getViewById(viewId);
+  updateView = action(
+    async (
+      workspaceSlug: string,
+      projectId: string,
+      viewId: string,
+      data: Partial<IProjectView>
+    ): Promise<IProjectView> => {
+      const currentView = this.getViewById(viewId);
 
-    const promiseRequests = [];
-    promiseRequests.push(this.viewService.patchView(workspaceSlug, projectId, viewId, data));
+      const promiseRequests = [];
+      promiseRequests.push(this.viewService.patchView(workspaceSlug, projectId, viewId, data));
 
-    runInAction(() => {
-      set(this.viewMap, [viewId], { ...currentView, ...data });
-    });
+      runInAction(() => {
+        set(this.viewMap, [viewId], { ...currentView, ...data });
+      });
 
-    if (data.access !== undefined && data.access !== currentView.access) {
-      promiseRequests.push(this.updateViewAccess(workspaceSlug, projectId, viewId, data.access));
+      const [response] = await Promise.all(promiseRequests);
+
+      return response;
     }
-
-    const [response] = await Promise.all(promiseRequests);
-
-    return response;
-  };
+  );
 
   /**
    * Deletes a view and removes it from the viewMap object
@@ -273,75 +248,6 @@ export class ProjectViewStore implements IProjectViewStore {
         if (this.rootStore.favorite.entityMap[viewId]) this.rootStore.favorite.removeFavoriteFromStore(viewId);
       });
     });
-  };
-
-  /** Locks view
-   * @param workspaceSlug
-   * @param projectId
-   * @param viewId
-   * @returns
-   */
-  lockView = async (workspaceSlug: string, projectId: string, viewId: string) => {
-    try {
-      const currentView = this.getViewById(viewId);
-      if (currentView?.is_locked) return;
-      runInAction(() => {
-        set(this.viewMap, [viewId, "is_locked"], true);
-      });
-      await this.viewService.lockView(workspaceSlug, projectId, viewId);
-    } catch (error) {
-      console.error("Failed to lock the view in view store", error);
-      runInAction(() => {
-        set(this.viewMap, [viewId, "is_locked"], false);
-      });
-    }
-  };
-
-  /**
-   * unlocks View
-   * @param workspaceSlug
-   * @param projectId
-   * @param viewId
-   * @returns
-   */
-  unLockView = async (workspaceSlug: string, projectId: string, viewId: string) => {
-    try {
-      const currentView = this.getViewById(viewId);
-      if (!currentView?.is_locked) return;
-      runInAction(() => {
-        set(this.viewMap, [viewId, "is_locked"], false);
-      });
-      await this.viewService.unLockView(workspaceSlug, projectId, viewId);
-    } catch (error) {
-      console.error("Failed to unlock view in view store", error);
-      runInAction(() => {
-        set(this.viewMap, [viewId, "is_locked"], true);
-      });
-    }
-  };
-
-  /**
-   * Updates View access
-   * @param workspaceSlug
-   * @param projectId
-   * @param viewId
-   * @param access
-   * @returns
-   */
-  updateViewAccess = async (workspaceSlug: string, projectId: string, viewId: string, access: EViewAccess) => {
-    const currentView = this.getViewById(viewId);
-    const currentAccess = currentView?.access;
-    try {
-      runInAction(() => {
-        set(this.viewMap, [viewId, "access"], access);
-      });
-      await this.viewService.updateViewAccess(workspaceSlug, projectId, viewId, access);
-    } catch (error) {
-      console.error("Failed to update Access for view", error);
-      runInAction(() => {
-        set(this.viewMap, [viewId, "access"], currentAccess);
-      });
-    }
   };
 
   /**
@@ -392,93 +298,6 @@ export class ProjectViewStore implements IProjectViewStore {
       runInAction(() => {
         set(this.viewMap, [viewId, "is_favorite"], true);
       });
-    }
-  };
-
-  /**
-   * Publishes View to the Public
-   * @param workspaceSlug
-   * @param projectId
-   * @param viewId
-   * @returns
-   */
-  publishView = async (workspaceSlug: string, projectId: string, viewId: string, data: TPublishViewSettings) => {
-    try {
-      const response = (await this.viewService.publishView(
-        workspaceSlug,
-        projectId,
-        viewId,
-        data
-      )) as TPublishViewDetails;
-      runInAction(() => {
-        set(this.viewMap, [viewId, "anchor"], response?.anchor);
-      });
-
-      return response;
-    } catch (error) {
-      console.error("Failed to publish view", error);
-    }
-  };
-
-  /**
-   * fetches Published Details
-   * @param workspaceSlug
-   * @param projectId
-   * @param viewId
-   * @returns
-   */
-  fetchPublishDetails = async (workspaceSlug: string, projectId: string, viewId: string) => {
-    try {
-      const response = (await this.viewService.getPublishDetails(
-        workspaceSlug,
-        projectId,
-        viewId
-      )) as TPublishViewDetails;
-      runInAction(() => {
-        set(this.viewMap, [viewId, "anchor"], response?.anchor);
-      });
-      return response;
-    } catch (error) {
-      console.error("Failed to fetch published view details", error);
-    }
-  };
-
-  /**
-   * updates already published view
-   * @param workspaceSlug
-   * @param projectId
-   * @param viewId
-   * @returns
-   */
-  updatePublishedView = async (
-    workspaceSlug: string,
-    projectId: string,
-    viewId: string,
-    data: Partial<TPublishViewSettings>
-  ) => {
-    try {
-      return await this.viewService.updatePublishedView(workspaceSlug, projectId, viewId, data);
-    } catch (error) {
-      console.error("Failed to update published view details", error);
-    }
-  };
-
-  /**
-   * un publishes the view
-   * @param workspaceSlug
-   * @param projectId
-   * @param viewId
-   * @returns
-   */
-  unPublishView = async (workspaceSlug: string, projectId: string, viewId: string) => {
-    try {
-      const response = await this.viewService.unPublishView(workspaceSlug, projectId, viewId);
-      runInAction(() => {
-        set(this.viewMap, [viewId, "anchor"], null);
-      });
-      return response;
-    } catch (error) {
-      console.error("Failed to unPublish view", error);
     }
   };
 }
