+++ apps/web/ee/services/project/index.ts
@@ -1,2 +1,1 @@
 export * from "./estimate.service";
-export * from "ce/services/project/view.service";
